Move MyButton under XAML dir
